aerospace: Ensure accordion layout is not used when restarting AeroSpace
Relevant comment from the maintainer on why this happens and the
suggested workaround:

When you launch AeroSpace and more than 3 windows get scheduled on the focused workspace
(e.g. you can use on-window-detected to schedule windows on different workspaces),
AeroSpace will switch the layout to accordion. It happens only on startup to
make onboarding experience better, because if there are too many windows open,
tiles layout doesn't look that good because of min-width window restriction,
and it may scare off new users. As a workaround, you can cancel the effect of
this "smart" behavior after-startup-command = ['layout tiles']